User Registration

+++ src/components/RegisterUser.js
@@ -2,16 +2,19 @@ import React from "react";
 
 import { axiosWithAuth } from "../utils/axiosWithAuth";
 
-class Login extends React.Component {
+class RegisterUser extends React.Component {
   state = {
+    error: "",
     credentials: {
       username: "",
-      password: ""
+      password: "",
+      passwordConfirm: ""
     }
   };
 
   handleChange = e => {
     this.setState({
+      ...this.state,
       credentials: {
         ...this.state.credentials,
         [e.target.name]: e.target.value
@@ -19,24 +22,36 @@ class Login extends React.Component {
     });
   };
 
-  login = e => {
+  register = e => {
     e.preventDefault();
-    axiosWithAuth()
-      .post("/login", this.state.credentials)
-      .then(res => {
-        localStorage.setItem("token", res.data.payload);
-        this.props.func();
-        this.props.history.push("/");
-      })
-      .catch(err => {
-        console.log("Err is: ", err);
-      });
+
+    if (this.state.credentials.password === this.state.credentials.passwordConfirm){
+        axiosWithAuth()
+        .post("/user", this.state.credentials)
+        .then(res => {
+          localStorage.setItem("token", res.data.payload);
+          this.props.func();
+          this.props.history.push("/");
+        })
+        .catch(err => {
+          console.log("Err is: ", err);
+          this.setState({
+            ...this.state,
+            error: err
+          })
+        });
+      } else {
+        this.setState({
+          ...this.state,
+          error: "Passwords do not match."
+        })
+      }
   };
 
   render() {
     return (
       <div className="loginForm">
-        <form onSubmit={this.login}>
+        <form onSubmit={this.register}>
           <h2>Please Log In</h2>
           <label htmlFor="username">Username: </label>
           <input
@@ -54,6 +69,15 @@ class Login extends React.Component {
             onChange={this.handleChange}
             placeholder="Password"
           /><br />
+          <label htmlFor="passwordConfirm">Confirm Password: </label>
+          <input
+            type="password"
+            name="passwordConfirm"
+            value={this.state.credentials.passwordConfirm}
+            onChange={this.handleChange}
+            placeholder="Confirm Password"
+          /><br />
+          {this.state.error !== ''?<div><p>{this.state.error}</p><br /></div>:<></>}
           <button className="loginBtn">Register</button>
         </form>
       </div>
@@ -61,4 +85,4 @@ class Login extends React.Component {
   }
 }
 
-export default Login;
+export default RegisterUser;
